proper shit
